src/back/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-decouple>=3.8
mssql-django>=1.4
pyodbc>=4.0.39